Move EsiClient from lib.rs into client.rs
